+++ src/common/mocks/index.ts
@@ -8,3 +8,4 @@ export * from './audio-features.mock'
 export * from './spotify-response.mock'
 export * from './image.mock'
 export * from './user.mock'
+export * from './axios-response.factory.mock'
