+++ src/common/utils/apply-authorization-header.util.ts
