+++ src/modules/statistics/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './audio-features-reducer'
-export * from './analysis-factory'
+export * from './audio-features.reducer'
+export * from './analysis.factory'
